+++ database/json.go
@@ -536,7 +536,7 @@ func execJSONArrInsert(db *DB, args [][]byte) redis.Reply {
 	}
 	entity, exists := db.GetEntity(key)
 	if !exists {
-		return protocol.MakeErrReply("ERR key does not exist")
+		return protocol.MakeErrReply("ERR could not perform this operation on a key that doesn't exist")
 	}
 	jv, ok := entity.Data.(*godisjson.JSONValue)
 	if !ok {
@@ -720,7 +720,7 @@ func execJSONToggle(db *DB, args [][]byte) redis.Reply {
 	}
 	entity, exists := db.GetEntity(key)
 	if !exists {
-		return &protocol.NullBulkReply{}
+		return protocol.MakeErrReply("ERR could not perform this operation on a key that doesn't exist")
 	}
 	jv, ok := entity.Data.(*godisjson.JSONValue)
 	if !ok {