+++ database/stream.go
@@ -954,7 +954,7 @@ func execXGroupSetID(db *DB, args [][]byte) redis.Reply {
 		return errReply
 	}
 	if s == nil {
-		return protocol.MakeErrReply("ERR no such key")
+		return protocol.MakeErrReply("ERR The XGROUP subcommand requires the key to exist. Note that for CREATE you may want to use the MKSTREAM option to create an empty stream automatically.")
 	}
 
 	group, err := s.GetGroup(groupName)